fix: add custom score calc

+++ backend/src/usecase/get_karaoke_score.py
@@ -7,7 +7,7 @@ class GetKaraokeScoreUseCase:
     def __init__(self, repository: KaraokeRepository):
         self.repository = repository
 
-    def execute(self, filename: str, stem: str, played_track: UploadFile) -> int:
+    def execute(self, filename: str, stem: str, played_track: UploadFile) -> float:
         if not stem.endswith('.wav'):
             stem = f'{stem}.wav'
         return self.repository.get_score(filename, stem, played_track)
